Scraper/requirements.txt
scrapy==2.12.0
psycopg2==2.9.10
python-dotenv==1.0.1
setuptools==59.6.0
wheel==0.45.1